Add test and fix bug in Student page

+++ database/migrations/2021_09_16_182217_create_tasks_table.php
@@ -16,7 +16,7 @@ class CreateTasksTable extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('id_degree');
+            $table->unsignedBigInteger('id_degree');
             $table->string('operation');
             $table->integer('pages');
             $table->integer('figure_1')->nullable();
